ci: add clang-tidy in build process
