+++ .github/workflows/ci.yml
@@ -28,25 +28,23 @@ jobs:
           check-path: ${{ matrix.path['check'] }}
           exclude-regex: ${{ matrix.path['exclude'] }}
 
-  cpp-linter:
+  linter:
+    name: Build And Run linter
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - uses: cpp-linter/cpp-linter-action@v2
-        id: linter
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v2
+      - name: Build linter image
+        uses: docker/build-push-action@v4
         with:
-          style: file
-          extensions: 'cpp,h'
-          version: 13
-          ignore: "endToEndTests|build|include/silo/roaring.h"
-          files-changed-only: false
-          tidy-checks: ''
-
-      - name: Fail fast!
-        if: steps.linter.outputs.checks-failed > 0
-        run: echo "Some files failed the linting checks!"
+          context: .
+          push: false
+          target: linter
+          tags: ${{ env.DOCKER_TEST_IMAGE_NAME }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+          file: Dockerfile_linter
 
   dockerImage:
     name: Build And Run Unit Tests
@@ -55,8 +53,7 @@ jobs:
       - uses: actions/checkout@v3
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v2
-      -
-        name: Login to GitHub Container Registry
+      - name: Login to GitHub Container Registry
         uses: docker/login-action@v2
         with:
           registry: ghcr.io
