+++ README.md
@@ -18,9 +18,8 @@ conan profile new myProfile --detect
 Insert info `os`, `os_build`, `arch` and `arch_build` of myProfile into `conanprofile.example` and rename
 to `conanprofile`.
 
-Build silo in `./build`. This build will loadDatabaseState and build the required libraries to `~/.conan/data/` (can not
-be set by hand).
-
+Build silo in `./build`. This build will load and build the required libraries to `~/.conan/data/` (can not be set by
+hand).
 ```shell
 ./build_with_conan
 ```
@@ -39,12 +38,27 @@ docker run -i silo
 ```
 
 # Testing
-For testing we use the framework [gtest](http://google.github.io/googletest/) and for mocking [gmock](http://google.github.io/googletest/gmock_cook_book.html). Tests are built using the same script as the production code: `./build_with_conan`.
 
-We use the convention, that each tested source file has its own test file, ending with `*.test.cpp`. The test file is placed in the same folder as the source file. If the function under test is described in a header file, the test file is located in the corresponding source folder. 
+For testing we use the framework [gtest](http://google.github.io/googletest/) and for
+mocking [gmock](http://google.github.io/googletest/gmock_cook_book.html). Tests are built using the same script as the
+production code: `./build_with_conan`.
 
+We use the convention, that each tested source file has its own test file, ending with `*.test.cpp`. The test file is
+placed in the same folder as the source file. If the function under test is described in a header file, the test file is
+located in the corresponding source folder.
 
 To run all tests, run
+
 ```shell
 build/silo_test
 ```
+
+For linting we use clang-tidy. The config is stored in `.clang-tidy`. It will run automatically with the build process
+and will throw errors accordingly. However it is rather slow. If you only want a fast build use
+
+```shell
+./build_with_conan build_without_clang_tidy
+```
+
+When pushing to github, a separate Docker image will be built, which runs the formatter. (This is a workaround, because
+building with clang-tidy under alpine was not possible yet.)
