+++ endToEndTests/test/queries/invalidMutationsMinProportion.json
@@ -7,8 +7,8 @@
     },
     "filter": {
       "type": "N-Of",
-      "n": 2,
-      "exactly": false,
+      "numberOfMatchers": 2,
+      "matchExactly": false,
       "children": []
     }
   },
