+++ src/database.test.cpp
@@ -6,8 +6,8 @@
 TEST(database_test, should_build_database_without_errors) {
    const silo::InputDirectory input_directory{"./testBaseData/"};
    const silo::OutputDirectory output_directory{"./build/"};
-   const silo::MetadataFilename metadata_filename{"minimal_set.tsv"};
-   const silo::SequenceFilename sequence_filename{"minimal_set.fasta"};
+   const silo::MetadataFilename metadata_filename{"minimal_metadata_set.tsv"};
+   const silo::SequenceFilename sequence_filename{"minimal_sequence_set.fasta"};
    auto config = silo::PreprocessingConfig(
       input_directory, output_directory, metadata_filename, sequence_filename
    );
